Review changes including refactoring Header, Packet etc into their own files
